Add index, 1st recipe, and links to 2 & 3 recipes
Create homepage for website, created and filled out the mushroom pasta
recipe page, and created files and links to the 2 new recipe pages:
holeinone breakfast and musclemommy yogurt.

In mushroom pasta, I created the template for what the future recipe
pages will hold, including the image, the description, the ingredients,
and the steps for the recipe.