Add 2 more recipes, edit index to include images
Fill out the holeinone recipe and musclemommy recipe pages. Include the
pictures of all the recipes on the index page to make it slightly more
appealing.

The new recipes follow the same format of the first recipe: picture,
description, ingredients, steps. I added a home page to the top to make
it easier to get back to the home page.